requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
facenet-pytorch==2.5.2
idna==3.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
